ez enemy states fix
- Turns out that all enemies in the vector point to the same state instance.
This means that once the first enemy switches state, it deletes that instance and causes all the enemies that come after it to have a pointer pointing to nothing (I think).
This is fine for Player cause there's only ever one of them.
The fix is to give each enemy a state instance of their own, which is done in enemyPool, where an initial state instance is initialized for each enemy in the vector.